requirements.txt
pygame
pyrebase4
requests
firebase_admin
python-dotenv